Added EF7 logging to console.  InMemoryBus no longer offloads handlers to new threads.  Instead it uses async await.  Yes, this will be slower, but it won't matter all that much under normal load.

+++ src/OxHack.Inventory.Query.Sqlite/Repositories/PhotoRepository.cs
@@ -8,38 +8,50 @@ using System.Threading.Tasks;
 
 namespace OxHack.Inventory.Query.Sqlite.Repositories
 {
-    public class PhotoRepository : IPhotoRepository
-    {
+	public class PhotoRepository : IPhotoRepository
+	{
 		private readonly DbContextOptions dbContextOptions;
+		private readonly SqliteWriteLock writeLock;
 
-		public PhotoRepository(DbContextOptions dbContextOptions)
+		public PhotoRepository(DbContextOptions dbContextOptions, SqliteWriteLock writeLock)
 		{
 			this.dbContextOptions = dbContextOptions;
+			this.writeLock = writeLock;
 		}
 
 		public async Task AddPhotoToItemAsync(Guid itemId, string photoFilename)
 		{
-			using (var dbContext = new InventoryDbContext(this.dbContextOptions))
+			lock (this.writeLock)
 			{
-				var newPhoto = new Models.Photo()
+				using (var dbContext = new InventoryDbContext(this.dbContextOptions))
 				{
-					Filename = photoFilename,
-					ItemId = itemId.ToString()
-				};
+					var newPhoto = new Models.Photo()
+					{
+						Filename = photoFilename,
+						ItemId = itemId.ToString()
+					};
 
-				dbContext.Photos.Add(newPhoto, GraphBehavior.SingleObject);
-				await dbContext.SaveChangesAsync();
+					dbContext.Photos.Add(newPhoto, GraphBehavior.SingleObject);
+					dbContext.SaveChanges();
+				}
 			}
 		}
 
 		public async Task RemovePhotoFromItemAsync(Guid itemId, string photoFilename)
 		{
-			using (var dbContext = new InventoryDbContext(this.dbContextOptions))
+			lock (this.writeLock)
 			{
-				var match = dbContext.Photos.Single(photo => photo.ItemId == itemId.ToString() && photo.Filename == photoFilename);
+				using (var dbContext = new InventoryDbContext(this.dbContextOptions))
+				{
+					var itemIdAsString = itemId.ToString();
+					var match = dbContext.Photos.SingleOrDefault(photo => photo.ItemId == itemIdAsString && photo.Filename == photoFilename);
 
-				dbContext.Photos.Remove(match);
-				await dbContext.SaveChangesAsync();
+					if (match != null)
+					{
+						dbContext.Photos.Remove(match);
+						dbContext.SaveChanges();
+					}
+				}
 			}
 		}
 	}
